pyproject.toml
[project]
name = "monger-manager"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
  {name = "Aaron Nicholas Gumapac", email = "aarn.gmpc@gmail.com"}
]
dependencies = [
  "pydantic>=2.9.2",
  "loguru>=0.7.2",
  "python-dotenv>=1.1.0",
  "aiohttp>=3.10.5",
  "polygon-api-client>=1.14.2",
  "ibapi @ file:///${PROJECT_ROOT}/lib/IBJts/source/pythonclient",
  "schedule>=1.2.2",
  "pytz>=2024.2",
  "numpy>=2.1.2",
  "pyyaml>=6.0.2",
  "anyio>=4.7.0",
  "clickhouse-connect>=0.8.16"
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.flake8]
ignore = [
  "E501"
]

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
  "ruff>=0.6.8",
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0"
]

[tool.pdm.scripts]
post_install = "pre-commit install"
lint = "pre-commit run --all-files"

[tool.ruff]
target-version = "py38"
line-length = 100
indent-width = 4
exclude = ["src/depricated"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
ignore = ["I001"]
select = ["E", "F", "I"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
force-single-line = false
force-sort-within-sections = true
lines-after-imports = 2
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E501"]
